+++ commands/music/play.js
@@ -17,16 +17,12 @@ module.exports = {
       voiceChannel: message.member.voice.channel.id,
       textChannel: message.channel.id,
     });
-    message.member.voice.channel.join().then((con) => {
-      //eslint-disable-next-line
-      con.voice.setDeaf(true).catch((err) => {});
-    });
-    player.connect();
     if (player.playing && player.queue.current) {
       player.queue.add(res.tracks[0]);
       message.react("👌");
       return message.reply(`Added \`${res.tracks[0].title}\` to Queue`);
     }
+    player.connect();
     // if (!player.playing && !player.paused && !player.queue.size) player.play();
     message.react("👌");
     player.play(res.tracks[0]);